images/requirements.txt
fastapi==0.111.1
httpx==0.27.2
pillow==10.4.0